🚚(rooms) split test file for rooms api
The file had become too long
