+++ tests/apps/core/api/test_core_api_frontend_configuration.py
