+++ tests/apps/core/api/test_core_api_groups.py
